Add custom exception handling to CrimeController

+++ src/main/java/se/iths/crimedatabase/controller/CrimeController.java
@@ -4,6 +4,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import se.iths.crimedatabase.entity.Crime;
+import se.iths.crimedatabase.exception.BadRequestException;
+import se.iths.crimedatabase.exception.NotFoundException;
 import se.iths.crimedatabase.service.CrimeService;
 
 import java.util.Optional;
@@ -20,17 +22,24 @@ public class CrimeController {
     @PostMapping
     public ResponseEntity<Crime> create(@RequestBody Crime crime) {
         Crime createdCrime = crimeService.create(crime);
+        if (crime.getName().isEmpty())
+            throw new BadRequestException("Crime name cannot be empty");
         return new ResponseEntity<>(createdCrime, HttpStatus.CREATED);
     }
 
     @DeleteMapping("{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        if (crimeService.findById(id).isEmpty())
+            throw new NotFoundException(responseMessage(id));
         crimeService.delete(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("{id}")
     public ResponseEntity<Optional<Crime>> findById(@PathVariable Long id) {
+        if (crimeService.findById(id).isEmpty())
+            throw new NotFoundException(responseMessage(id));
+
         Optional<Crime> foundCrime = crimeService.findById(id);
         return new ResponseEntity<>(foundCrime, HttpStatus.OK);
     }
@@ -38,7 +47,14 @@ public class CrimeController {
     @GetMapping
     public ResponseEntity<Iterable<Crime>> findAll() {
         Iterable<Crime> crimes = crimeService.findAll();
+        if (!crimes.iterator().hasNext())
+            throw new NotFoundException("No crimes found");
+
         return new ResponseEntity<>(crimes, HttpStatus.OK);
+    }
+
+    private String responseMessage(Long id) {
+        return "Crime with id: " + id + " cannot be found";
     }
 
 }
